split emails by number of attachments
AWS SES is only capable of sending 10MB per email
Hence when there are many attachments there is the possibility this
threshold is passed
To workaround this the emails are split based on the number of
attachments
As attachments are Base64 encoded there is an overhead and once
attachments reach ~7.5MB there is risk the message will become too large
