+++ spec/spec_helper.rb
@@ -107,7 +107,8 @@ SimpleCov.start do
   add_filter '/config/'
   add_filter '/spec/'
 end
-SimpleCov.minimum_coverage 90
+
+SimpleCov.minimum_coverage 91
 SimpleCov.formatter = SimpleCov::Formatter::MultiFormatter.new([
   SimpleCov::Formatter::HTMLFormatter,
   SimpleCov::Formatter::Console
